create function doMethod to know if the method return modelview or string or value or value unknown
