+++ src/mg/noobframework/utils/MviewController.java
@@ -1,0 +1,5 @@
+package mg.noobframework.utils;
+
+public class MviewController {
+    
+}
